# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: PICA API
  version: 0.2.0
  description: |
    Backend `pica-back`, módulo de usuarios (registro, sesión, perfil, backoffice).

    **Convenciones**
    - Todo bajo `/api/v1`. `Authorization: Bearer <access token>` salvo lo marcado `security: []`.
    - Errores en `application/problem+json` (RFC 9457) con un campo `codigo` (ver `ProblemDetail`).
    - Para no repetirlos en cada operación: todo endpoint autenticado puede responder
      **401** (`NO_AUTENTICADO`, `TOKEN_INVALIDO`, `TOKEN_VENCIDO`), todo `/admin` **403**
      `SIN_PERMISO`, todo `{id}` **404** (`USUARIO_NO_ENCONTRADO`, `PERSONA_NO_ENCONTRADA`,
      `ROL_NO_ENCONTRADO`), y todo request con body **400** `VALIDACION`.
      Abajo solo figuran los errores con código propio de cada operación.
    - Fechas ISO-8601: `date` = `YYYY-MM-DD`, `date-time` = UTC.
    - Baja lógica en todo: lo eliminado no sale en listados salvo `incluirEliminados=true`,
      y viene con `eliminado: true`.
servers:
  - {url: 'http://localhost:8080', description: Local}
  - {url: 'https://PENDIENTE.up.railway.app', description: Staging (completar cuando esté T3)}
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/roles:
    get:
      tags: [Admin – Roles y permisos]
      summary: Listar roles
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_VER
      description: >-
        Paginado igual que los otros listados para reusar la tabla genérica. `cantidadUsuarios` cuenta
        usuarios no eliminados. `q` busca en `nombre` y `nombreAmigable`. Se puede ordenar por `id`
        (default), `nombre`, `nombreAmigable`, `estado` y `creadoEn`; otro campo da 400 `VALIDACION`.
      parameters:
        - {$ref: '#/components/parameters/q'}
        - {name: estado, in: query, schema: {$ref: '#/components/schemas/Estado'}}
        - {$ref: '#/components/parameters/incluirEliminados'}
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/size'}
        - {$ref: '#/components/parameters/sort'}
      responses:
        '200': {description: Página, content: {application/json: {schema: {$ref: '#/components/schemas/PageRolResumen'}}}}
    post:
      tags: [Admin – Roles y permisos]
      summary: Crear un rol
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_CREAR
      description: Nace sin permisos; se cargan con `PUT /admin/roles/{id}/permisos`. 409 `ROL_DUPLICADO`.
      requestBody: {$ref: '#/components/requestBodies/Rol'}
      responses:
        '201': {$ref: '#/components/responses/Rol'}
        '409': {$ref: '#/components/responses/Conflicto'}
  /api/v1/admin/roles/{id}:
    parameters: [{$ref: '#/components/parameters/id'}]
    get:
      tags: [Admin – Roles y permisos]
      summary: Ver un rol (con permisos)
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_VER
      description: >-
        Incluye los dados de baja (vienen con `eliminado: true`) para poder abrirlos desde el listado
        y reactivarlos.
      responses:
        '200': {$ref: '#/components/responses/Rol'}
    put:
      tags: [Admin – Roles y permisos]
      summary: Modificar un rol
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_EDITAR
      description: >-
        Si no viene `estado` se mantiene el que tenía. 403 `ROL_PROTEGIDO`: SUPER_USUARIO siempre;
        PARTICIPANTE si cambia `nombre` o `estado` (el registro se lo asigna a todo usuario nuevo),
        el nombre amigable y la descripción sí se cambian. 404 si está dado de baja (primero se reactiva).
        409 `ROL_DUPLICADO`, también contra roles dados de baja.
      requestBody: {$ref: '#/components/requestBodies/Rol'}
      responses:
        '200': {$ref: '#/components/responses/Rol'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Conflicto'}
    delete:
      tags: [Admin – Roles y permisos]
      summary: Baja lógica
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_ELIMINAR
      description: >-
        Un rol eliminado no se puede asignar y deja de sumar permisos (solo cuentan roles activos).
        Las asignaciones se conservan: al reactivarlo sus usuarios lo recuperan.
        403 `ROL_PROTEGIDO` (SUPER_USUARIO y PARTICIPANTE). 404 si ya estaba dado de baja.
      responses:
        '204': {description: Dado de baja}
        '403': {$ref: '#/components/responses/Error'}
  /api/v1/admin/roles/{id}/reactivar:
    post:
      tags: [Admin – Roles y permisos]
      summary: Reactivar
      x-jira: PICA-125
      x-owner: Tomás
      x-permiso: ROL_ELIMINAR
      description: >-
        No está en la subtarea; se agrega para que roles se comporte igual que usuarios y personas en la tabla.
        Sobre un rol que no estaba dado de baja no hace nada y lo devuelve.
      parameters: [{$ref: '#/components/parameters/id'}]
      responses:
        '200': {$ref: '#/components/responses/Rol'}
  /api/v1/admin/roles/{id}/permisos:
    put:
      tags: [Admin – Roles y permisos]
      summary: Reemplazar los permisos de un rol
      x-jira: PICA-126
      x-owner: Tomás
      x-permiso: ROL_EDITAR
      description: >-
        Lista completa, reemplaza (vacía = sin permisos; los repetidos cuentan una vez). Los usuarios
        lo ven en su próximo token (los permisos se calculan al emitir el JWT).
        403 `ROL_PROTEGIDO` (SUPER_USUARIO; a PARTICIPANTE sí se le cambian). 404 si el rol está dado
        de baja, y `PERMISO_NO_ENCONTRADO` con los códigos que no existen en `invalidos` (no se guarda nada).
        409 `ROL_INACTIVO`.
      parameters: [{$ref: '#/components/parameters/id'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [permisos], properties: {permisos: {type: array, items: {$ref: '#/components/schemas/CodigoPermiso'}}}}}}
      responses:
        '200': {$ref: '#/components/responses/Rol'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Conflicto'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        RS256, 15 min. Claims: `sub` (id de usuario), `username`, `roles` (nombres técnicos activos),
        `permisos` (unión de los permisos de sus roles activos), `jti`, `iat`, `exp`.
        El filtro carga las authorities desde `permisos`; los `@PreAuthorize` usan `hasAuthority('USUARIO_VER')`.
  parameters:
    id: {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
    q: {name: q, in: query, schema: {type: string, maxLength: 100}}
    incluirEliminados: {name: incluirEliminados, in: query, schema: {type: boolean, default: false}}
    page: {name: page, in: query, description: Desde 0, schema: {type: integer, minimum: 0, default: 0}}
    size: {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    sort:
      name: sort
      in: query
      description: '`campo,asc` o `campo,desc`; se puede repetir. Ej: `sort=apellidos,asc&sort=nombres,asc`'
      schema: {type: array, items: {type: string}}
      style: form
      explode: true
  requestBodies:
...
    Rol: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RolRequest'}}}}
  responses:
...
    Rol: {description: Rol con permisos, content: {application/json: {schema: {$ref: '#/components/schemas/RolDetalle'}}}}
    Error: {description: Ver códigos en la descripción de la operación, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}}
    Conflicto: {description: Regla de negocio que no se cumple; el código dice cuál, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}}
  schemas:
...
    Id: {type: integer, format: int64, minimum: 1}
...
    Nombre: {type: string, minLength: 1, maxLength: 100}
...
    Estado: {type: string, enum: [ACTIVO, INACTIVO], description: Persona y rol. La baja lógica es aparte (`eliminado`).}
...
    CodigoPermiso:
      type: string
      description: Seed de V4 (PICA-124). Si aparece uno nuevo se agrega acá.
      enum: [USUARIO_VER, USUARIO_CREAR, USUARIO_EDITAR, USUARIO_ELIMINAR,
             PERSONA_VER, PERSONA_CREAR, PERSONA_EDITAR, PERSONA_ELIMINAR,
             ROL_VER, ROL_CREAR, ROL_EDITAR, ROL_ELIMINAR, ROL_ASIGNAR]

    # --- errores -------------------------------------------------------------
    CodigoError:
      type: string
      description: |
        Todos los códigos de la API; el front mapea cada uno a un mensaje. Quien agrega uno en el
        backend lo agrega acá.
        Genéricos: VALIDACION 400 · NO_AUTENTICADO / TOKEN_INVALIDO / TOKEN_VENCIDO 401 · SIN_PERMISO 403 · *_NO_ENCONTRADO 404.
        Auth: CREDENCIALES_INVALIDAS 401 · EMAIL_NO_VERIFICADO / USUARIO_BLOQUEADO 403 ·
        REFRESH_INVALIDO / REFRESH_REUTILIZADO 401 · TOKEN_EXPIRADO / TOKEN_USADO / CODIGO_INVALIDO / PASSWORD_ACTUAL_INCORRECTA 400.
        Negocio (409): USERNAME_DUPLICADO · EMAIL_DUPLICADO · DOCUMENTO_DUPLICADO · ROL_DUPLICADO ·
        PERSONA_INACTIVA · PERSONA_CON_USUARIO · DOCUMENTO_NO_EDITABLE · ROL_INACTIVO · USUARIO_NO_ACTIVO.
        Protecciones (403): USUARIO_PROTEGIDO (Admin del sistema) · ROL_PROTEGIDO (SUPER_USUARIO; PARTICIPANTE en nombre, estado y baja) · ULTIMO_ASIGNADOR.
      enum: [VALIDACION, NO_AUTENTICADO, TOKEN_INVALIDO, TOKEN_VENCIDO, SIN_PERMISO, ERROR_INTERNO,
             USUARIO_NO_ENCONTRADO, PERSONA_NO_ENCONTRADA, ROL_NO_ENCONTRADO, PERMISO_NO_ENCONTRADO,
             CREDENCIALES_INVALIDAS, EMAIL_NO_VERIFICADO, USUARIO_BLOQUEADO, REFRESH_INVALIDO, REFRESH_REUTILIZADO,
             TOKEN_EXPIRADO, TOKEN_USADO, CODIGO_INVALIDO, PASSWORD_ACTUAL_INCORRECTA,
             USERNAME_DUPLICADO, EMAIL_DUPLICADO, DOCUMENTO_DUPLICADO, ROL_DUPLICADO,
             PERSONA_INACTIVA, PERSONA_CON_USUARIO, DOCUMENTO_NO_EDITABLE, ROL_INACTIVO, USUARIO_NO_ACTIVO,
             USUARIO_PROTEGIDO, ROL_PROTEGIDO, ULTIMO_ASIGNADOR]
    ErrorCampo:
      type: object
      required: [campo, codigo]
      description: >-
        Un ítem de `errores` en los 400. `codigo` sale de la anotación que falló: @NotBlank/@NotNull →
        REQUERIDO, @Email/@Pattern → FORMATO_INVALIDO, @Size → LONGITUD, @PastOrPresent → FECHA_FUTURA,
        política de clave → PASSWORD_DEBIL, otra → VALOR_INVALIDO.
      properties:
        campo: {type: string, description: Nombre del campo tal como viene en el JSON}
        codigo: {type: string, enum: [REQUERIDO, FORMATO_INVALIDO, LONGITUD, FECHA_FUTURA, PASSWORD_DEBIL, VALOR_INVALIDO]}
        mensaje: {type: string, description: Orientativo; el front muestra el suyo según `codigo`}
    ProblemDetail:
      type: object
      required: [status, codigo]
      description: >-
        El ProblemDetail de Spring (RFC 9457) más `codigo` (siempre) y `errores` (solo en VALIDACION).
        Lo arma un @RestControllerAdvice en common/error; ningún controller devuelve errores a mano.
      properties:
        type: {type: string, default: 'about:blank'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        codigo: {$ref: '#/components/schemas/CodigoError'}
        errores: {type: array, items: {$ref: '#/components/schemas/ErrorCampo'}}
        invalidos: {type: array, items: {type: string}, description: 'Solo en PERMISO_NO_ENCONTRADO: los códigos que no existen'}
      examples:
        - {type: 'about:blank', title: Bad Request, status: 400, detail: Hay campos inválidos, instance: /api/v1/auth/registro,
         codigo: VALIDACION, errores: [{campo: password, codigo: PASSWORD_DEBIL, mensaje: 'Mínimo 8, una mayúscula y un número'}]}

    # --- paginación (forma de Spring con serialization-mode=via_dto) ---------
    PageInfo:
      type: object
      required: [size, number, totalElements, totalPages]
      properties:
        size: {type: integer}
        number: {type: integer, description: 'Página actual, desde 0'}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
    # OpenAPI no tiene genéricos, así que hay un Page por tipo; todos tienen la misma forma.
...
    PageRolResumen:
      type: object
      required: [content, page]
      properties: {content: {type: array, items: {$ref: '#/components/schemas/RolResumen'}}, page: {$ref: '#/components/schemas/PageInfo'}}

    # --- auth ----------------------------------------------------------------
...
    RolResumen:
      type: object
      required: [id, nombre, nombreAmigable, estado, esSistema, eliminado, cantidadUsuarios]
      properties:
        id: {$ref: '#/components/schemas/Id'}
        nombre: {type: string}
        nombreAmigable: {type: string}
        descripcion: {type: [string, 'null']}
        estado: {$ref: '#/components/schemas/Estado'}
        esSistema: {type: boolean, description: true solo para SUPER_USUARIO; solo lectura en la pantalla}
        eliminado: {type: boolean}
        cantidadUsuarios: {type: integer}
    RolDetalle:
      allOf:
        - {$ref: '#/components/schemas/RolResumen'}
        - type: object
          required: [permisos, creadoEn]
          properties:
            permisos: {type: array, items: {$ref: '#/components/schemas/CodigoPermiso'}}
            eliminadoEn: {type: [string, 'null'], format: date-time}
            creadoEn: {type: string, format: date-time}
            modificadoEn: {type: [string, 'null'], format: date-time}
    RolRequest:
      type: object
      required: [nombre, nombreAmigable]
      properties:
        nombre:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,49}$'
          examples: [VEEDOR]
          description: Técnico, mayúsculas y guión bajo. Es lo que va en el claim `roles`; no se renombra sin avisar.
        nombreAmigable: {$ref: '#/components/schemas/Nombre'}
        descripcion: {type: [string, 'null'], maxLength: 500}
        estado:
          allOf: [{$ref: '#/components/schemas/Estado'}]
          description: En el alta, ACTIVO si no viene; en la modificación, si no viene se mantiene.
...